Infection script EventManager entfernt und eigener timer hinzugefügt
Opalma tree aus listen löschen entfernt
bool hinzugefügt
NodeManager debug gelöscht

+++ Assets/Scripts/ShroomTree/OpalmaCareSystem.cs
@@ -7,9 +7,11 @@ public class OpalmaCareSystem : Singleton<OpalmaCareSystem> {
 
 
     //List with all infected trees
+	[SerializeField]
 	List<ShroomTree> infectedTrees = new List<ShroomTree>();
 
     //List with all infected trees which get dmg over time
+	[SerializeField]
 	List<ShroomTree> dyingTrees= new List<ShroomTree>();
 
     //Pick value
@@ -34,7 +36,6 @@ public class OpalmaCareSystem : Singleton<OpalmaCareSystem> {
     public void addInfectedTree(ShroomTree tree)
     {
         tree.setInfection(true);
-
         infectedTrees.Add(tree);
     }
 
@@ -48,12 +49,17 @@ public class OpalmaCareSystem : Singleton<OpalmaCareSystem> {
 
     public void handleDmg()
     {
-        foreach (ShroomTree tree in dyingTrees)
-        {   if (tree.getHP()-dmg > .0f)
-                tree.dealDamage(dmg);
-            else
-                tree.killTree();
-        }
+	 	foreach (ShroomTree tree in dyingTrees)
+	    {   
+			if (tree.alive) {
+				if (tree.getHP () - dmg > .0f) {
+					tree.dealDamage (dmg);
+				} else {
+					tree.alive = false;
+					tree.killTree ();
+				}
+			}
+	   	}
     }
 
 
@@ -64,9 +70,9 @@ public class OpalmaCareSystem : Singleton<OpalmaCareSystem> {
         foreach(ShroomTree tree in infectedTrees)
         {
             tree.incrIntegrity();
-            if(tree.getIntegrity() >= tree.getMaxIntegrity())
+			if(tree.getIntegrity() >= tree.getMaxIntegrity() && !tree.dies)
             {
-				infectedTrees.Remove (tree);
+				tree.dies = true;
 				addDyingTree(tree);
 				tree.gameObject.AddComponent<Infection> ();
             }
